feat, test: strategies implement unimind interface
